Add decisions.hpp + code cleanup
